+++ CHANGELOG.md
@@ -7,13 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-### Added
-
-- Docs/CI: sphinx linkcheck job and tox environment/command (`tox -e links`) (#255, @ugognw)
-- Docs: add carbonate system tutorial (#204, @NikhilDhruv)
-
 ### Fixed
 
+- `Solution` engine, solvent, database were not inherited by the sum of `Solution` objects (#258, ugognw)
+- Bugs where incorrect $\alpha_1$ and $\alpha_2$ parameters were used to calculate activity
+  coefficients and solute molar volumes for salts with divalent (or greater) ions (#258, ugognw)
+- calculation of salt concentrations in `Solution.get_salt_dict` for salts containing polyvalent cations (#258, ugognw)
+- `Solution.get_salt_dict` now respects the `cutoff` parameter (#258, ugognw)
+  - Note that `cutoff` is interpreted in units of moles per kilogram of solution (#258, ugognw)
+- `Solution.get_salt_dict` always returns a salt dictionary sorted in order of decreasing salt concentration (#258, ugognw)
 - `Solution.__init__`: Raise `ValueError` if a user sets inconsistent `H[+1]` in `solutes` and
   `pH` keyword arguments (#270, @gnuhpdiem, @rkingsbury). Previously, if the user set `H[+1]` in `solutes`, it's value would silently override the `pH` kwarg. Now, you will get a `ValueError`
   if the two are inconsistent, unless the `pH` kwarg is kept at the default value. In that case,
@@ -29,8 +31,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docs: Minor fixes for private / cached methods (#197, @githubalexliu)
 - Docs: Edit documentation of `debye_parameter_B` (#196, @YitongPan1)
 
+### Added
+
+- Docs/CI: sphinx linkcheck job and tox environment/command (`tox -e links`) (#255, @ugognw)
+- Docs: add carbonate system tutorial (#204, @NikhilDhruv)
+
 ### Changed
 
+- **BREAKING** - the return value of `Solution.get_salt_dict` now includes `Salt` objects instead of keys corresponding
+  to `cation` and `anion`. See the example in the docstring for how to adapt existing code to accommodate this
+  change. (#258, @ugognw)
+- **BREAKING** - `Solution.get_salt_dict` no longer returns an entry for water (#258)
+- **BREAKING** - `Solution.get_salt` will not return water and may return `None` if no salt is present.
+  Previously, `Solution.get_salt` would have returned a `Salt` representing water. (#258)
 - Ensure more consistent column formatting in `Solution.print()` (#269, @rkingsbury)
 - Switch `math.log10` to `np.log10` in `Solution.p()` (#269, @rkingsbury)
 - update pre-commit configuration (#269, @rkingsbury)
@@ -43,6 +56,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docs: `tox -e docs` command configured to fail on warning (#255, ugognw)
 - Docs: ReadTheDocs built with Python 3.11 (#255, ugognw)
 - Use `importlib` to locate test files (#241, @SuixiongTay)
+- Support `numpy>2.0`
+- Bump `pint` to `0.24.4` for `numpy` `v2.0` compatibility and to mitigate CI issues (#239, @SuixiongTay, @rkingsbury)
+- CI: add `python` `v3.13` to post-merge unit tests
 - bump `pymatgen` to `v2025.1.9`
 - bump `maggma` to `v0.71.4`
 
@@ -53,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Solution.add_solvent` has been marked for deprecation and will be removed in a future
   release. Use `add_solute` instead.
 - Python 3.9 version classifier in pyproject.toml (#247, @ugognw)
+- `Solution.list_salts` (use `Solution.get_salt_dict()` instead) (#258)
 
 ## [1.2.0] - 2024-09-24